% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASGSmaps-package.R
\docType{package}
\name{ASGSmaps-package}
\alias{ASGSmaps}
\alias{ASGSmaps-package}
\title{ASGSmaps: An API Wrapper for Accessing Australian Statistical Geography Standard Data from the ABS}
\description{
Functions for accessing the Australian Statistical Geography Standard (ASGS) Edition 3 API used by the Australia Bureau of Statistics. Allows easy access and use of geographical files from the ABS in R.
}
\author{
\strong{Maintainer}: Melanie Renkin \email{mrenkin@gmail.com}

}
\keyword{internal}
