% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_functions.R
\name{available_services}
\alias{available_services}
\title{Show Geographical Services Available}
\usage{
available_services()
}
\value{
A character vector of geographic services available
}
\description{
Show Geographical Services Available
}
\examples{
available_services()
}
