% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_functions.R
\name{available_standards}
\alias{available_standards}
\title{Show Geographical Standards Available}
\usage{
available_standards(service)
}
\arguments{
\item{service}{A character vector from the list of
\code{\link[ASGSmaps]{available_services}}}
}
\value{
A list of geographic standards available
}
\description{
Show Geographical Standards Available
}
\examples{
available_standards(service = "ASGS2021")

}
