% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_functions.R
\name{field_values}
\alias{field_values}
\title{Show Values in Specified Standard Fields}
\usage{
field_values(service = NULL, standard = NULL, field = NULL, unique = TRUE)
}
\arguments{
\item{service}{A character vector with one element from the list of
\code{\link[ASGSmaps:available_services]{available_services}}}

\item{standard}{A character vector with one element from the list of
\code{\link[ASGSmaps:available_standards]{available_standards}}}

\item{field}{A character vector with one element from the list of
\code{\link[ASGSmaps:available_standards]{standard_fields}}}

\item{unique}{Boolean operator either TRUE or FALSE}
}
\value{
A data frame of the values in each field queried.
}
\description{
Show Values in Specified Standard Fields
}
\examples{
# get all of the values in the field 'STATE_NAME_2021' in the 'STE' standard
# in the 'ASGS2021' service

values <- field_values(
  service = "ASGS2021",
  standard = "STE",
  field = "STATE_NAME_2021"
)

# get all of the values in the fields 'STATE_NAME_2021' and 'STATE_CODE_2021'
# in the 'STE' standard in the 'ASGS2021' service

values <- field_values(
  service = "ASGS2021",
  standard = "STE",
  field = "STATE_NAME_2021, STATE_CODE_2021"
)

# get all of the values in all of the fields in the 'STE' standard in the
# 'ASGS2021' service

values <- field_values(
  service = "ASGS2021",
  standard = "STE",
  field = "*"
)
}
