% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_functions.R
\name{standard_description}
\alias{standard_description}
\title{Show Descriptions of Geographical Standards}
\usage{
standard_description(service, standard)
}
\arguments{
\item{service}{A string from the list of
\code{\link[ASGSmaps:available_services]{available_services}}}

\item{standard}{A string from the list of
\code{\link[ASGSmaps:available_standards]{available_standards}}}
}
\value{
A string containing a description of the standard
}
\description{
Show Descriptions of Geographical Standards
}
\examples{
standard_description(
  service = "ASGS2021",
  standard = "LGA"
)
}
