% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_functions.R
\name{standard_fields}
\alias{standard_fields}
\title{Show Search fields for Geographical Standards}
\usage{
standard_fields(service, standard)
}
\arguments{
\item{service}{A character vector with one element from the list of
\code{\link[ASGSmaps:available_services]{available_services}}}

\item{standard}{A character vector with one element from the list of
\code{\link[ASGSmaps:available_standards]{available_standards}}}
}
\value{
A list of fields available to search in the query
}
\description{
Show Search fields for Geographical Standards
}
\examples{
standard_fields(
  service = "ASGS2021",
  standard = "LGA"
)
}
